Add validations

+++ src/Controllers/product.controller.ts
@@ -1,4 +1,5 @@
 import {  Request, Response} from "express";
+import { validationResult } from "express-validator";
 import { productService } from "../Services/product.service";
 
 const _productService = new productService();
@@ -10,21 +11,41 @@ export class productController{
         else res.status(500).json(response);
    }
     async getById(req: Request,res: Response){
+        const errors = validationResult(req); 
+        if (!errors.isEmpty()) {
+            res.status(400).json({ error: errors.array() });
+            return;
+        }
         const response = await _productService.getById(req);
         if (response.success) res.status(200).json(response);
         else res.status(404).json(response);
     }
     async createProduct(req: Request,res: Response){
+        const errors = validationResult(req); 
+        if (!errors.isEmpty()) {
+            res.status(400).json({ error: errors.array() });
+            return;
+        }
         const response = await _productService.createProduct(req);
         if (response.success) res.status(201).json(response);
         else res.status(500).json(response);
     }
     async updateProduct(req: Request,res: Response){
+        const errors = validationResult(req); 
+        if (!errors.isEmpty()) {
+            res.status(400).json({ error: errors.array() });
+            return;
+        }
         const response = await _productService.updateProduct(req);
         if (response.success) res.status(200).json(response);
         else res.status(500).json(response);
     }
     async deleteProduct(req: Request, res: Response){
+        const errors = validationResult(req); 
+        if (!errors.isEmpty()) {
+            res.status(400).json({ error: errors.array() });
+            return;
+        }
         const response = await _productService.deleteProduct(req);
         if (response.success) res.status(200).json(response);
         else res.status(500).json(response);
